+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@react-three/drei": "^9.56.14",
     "@react-three/fiber": "^8.18.0",
+    "@react-three/rapier": "^1.0.0",
     "@sanity/image-url": "^1.1.0",
     "@types/three": "^0.176.0",
     "axios": "^1.8.2",
@@ -25,6 +26,7 @@
     "next-sanity": "^9.10.1",
     "react": "^18",
     "react-dom": "^18",
+    "react-is": "^19.1.0",
     "swr": "^2.3.3"
   },
   "devDependencies": {
